+++ deploy/7-dependencies-pendle-oracle.ts
@@ -27,5 +27,6 @@ const func: DeployFunction = async ({ getNamedAccounts, deployments, network }: 
 };
 
 export default func;
-func.tags = ["pendle"];
-func.skip = async (hre: HardhatRuntimeEnvironment) => hre.network.name !== "sepolia";
+func.tags = ["pendle", "mock-pt-oracle"];
+func.skip = async (hre: HardhatRuntimeEnvironment) =>
+  hre.network.name !== "sepolia" && hre.network.name !== "bsctestnet";